worker_kb/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kb-worker"
version = "0.1.0"
description = "ETL / Indexer / Scheduler worker for local multimodal knowledge base"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "OpenAI"}]
dependencies = [
  "pydantic>=2.8,<3",
  "pydantic-settings>=2.3,<3",
  "psycopg[binary]>=3.2,<4",
  "neo4j>=5.28,<6",
  "httpx>=0.27,<1",
  "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-cov>=5,<6",
  "ruff>=0.6,<1",
  "mypy>=1.11,<2",
]

[project.scripts]
kb-worker = "kb_worker.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
